pyproject.toml
[tool.poetry]
name = "pytorch-ssdir"
version = "0.1.0"
description = "Single Shot Multi-Box Detect Infer & Repeat implementation in PyTorch"
authors = ["Piotr Zieliński <piotrekzie100@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-ssd = "^1.7.0"
pytorch-lightning = "1.1.8"
torchvision = "^0.8.2"
pyro-ppl = "^1.5.2"
wandb = "^0.10.12"
numpy = "^1.19.3"
pillow = "^8.0.1"
networkx = "^2.5"
pywavelets = "^1.1.1"
tifffile = "^2020.10.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-cov = "^2.10.0"
pytest-xdist = "^2.1.0"
ipython = "^7.19.0"

[tool.poetry.scripts]
ssdir = "pytorch_ssdir.main:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.flakehell]
max-line-length = 88
ignore = ["E203", "E231"]

[pytest]
addopts = "--color=yes"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = "True"
